fix for long passwords
i ran into a little bug, where the password obviously exceeded the linewrap threshold, as i am tryiing to follow NIST guidelines on passwords.
this resulted in a line wrap in the middle of the encoded string, thus invalidating the password. the chomp only removes the whitespaces before an after. strict mode does help with not even adding line feeds.